Refactoring.
Isolated the logic to find a new crash report. The rationale is that this code is non-trivial so having a standalone C API makes the flow more clear (no semi-global state in the form of instance variables) and will also allow for better testing.

I did not change the functionality of the code, though some possible improvements:

 1. Do not search the pre-SnowLeopard folder for crash reports when running on SnowLeopard or better (if the SnowLeopard folder is empty, we may present the user with a really old crash report from pre-SnowLeopard).
 2. Do not show crash reports which are older than e.g. a day. If an old report is found, it’s likely because the user deleted his defaults. This is somewhat remedied by SnowLeopard cleaning up the Diagnostics Reports folder on a biweekly basis.