Use non-localized name for submit and log search.
I haven’t actually seen anyone localize the CFBundleExecutable key, but regardless, it doesn’t make sense to base the application name on the user’s locale when submitting a crash report.

Likewise the system.log is most likely going to use the process name (although there seems to be exceptions, maybe programs which log output via asl?), so here we shouldn’t use a localized name either.

The only place we want localized names is in the UI and a previous commit changed the UI to use the (proper) localized application name.